reorganized the storage package
