+++ src/main/java/astminer/examples/AllJavaFiles.java
@@ -3,6 +3,9 @@ package astminer.examples;
 import astminer.common.model.*;
 import astminer.parse.java.GumTreeJavaParser;
 import astminer.storage.*;
+import astminer.storage.path.CsvPathStorage;
+import astminer.storage.path.PathBasedStorage;
+import astminer.storage.path.PathBasedStorageConfig;
 
 import java.io.FileInputStream;
 import java.io.IOException;
