+++ src/main/kotlin/astminer/cli/ProjectParser.kt
@@ -1,7 +1,7 @@
 package astminer.cli
 
-import astminer.storage.CsvAstStorage
-import astminer.storage.DotAstStorage
+import astminer.storage.ast.CsvAstStorage
+import astminer.storage.ast.DotAstStorage
 import astminer.common.getProjectFilesWithExtension
 import astminer.common.preOrder
 import astminer.storage.Storage
