+++ src/main/kotlin/astminer/examples/AllJavaMethods.kt
@@ -5,6 +5,8 @@ import astminer.parse.java.GumTreeJavaNode
 import astminer.parse.java.GumTreeJavaParser
 import astminer.parse.java.GumTreeJavaMethodSplitter
 import astminer.storage.*
+import astminer.storage.path.CsvPathStorage
+import astminer.storage.path.PathBasedStorageConfig
 import java.io.File
