+++ src/main/kotlin/astminer/storage/path/CsvPathStorage.kt
@@ -1,6 +1,7 @@
-package astminer.storage
+package astminer.storage.path
 
 import astminer.common.model.PathContextId
+import astminer.storage.TokenProcessor
 
 class CsvPathStorage(
     outputDirectoryPath: String,
